BSERV-19054 Re-enable plugin upload for test Helm chart
Modern Bitbucket no longer permits plugin upload via UPM by default.
This is however useful for testing and in test and development
environments we've re-enabled the functionality, for example when the
product is run in dev-mode, or started with AMPS. The same should be
done for the test helm charts.